Start moving over some string tests
