+++ python/cudf/cudf/tests/test_string.py
@@ -1,12 +1,9 @@
 # Copyright (c) 2018-2025, NVIDIA CORPORATION.
 
-from decimal import Decimal
 from sys import getsizeof
 
-import cupy
 import numpy as np
 import pandas as pd
-import pyarrow as pa
 import pytest
 
 import rmm
@@ -17,8 +14,6 @@ from cudf.core.column.string import StringColumn
 from cudf.core.index import Index
 from cudf.testing import assert_eq
 from cudf.testing._utils import (
-    DATETIME_TYPES,
-    NUMERIC_TYPES,
     assert_exceptions_equal,
 )
 from cudf.utils import dtypes as dtypeutils
@@ -48,290 +43,6 @@ def ps_gs(data, index):
     ps = pd.Series(data, index=index, dtype="str", name="nice name")
     gs = cudf.Series(data, index=index, dtype="str", name="nice name")
     return (ps, gs)
-
-
-@pytest.mark.parametrize("construct", [list, np.array, pd.Series, pa.array])
-def test_string_ingest(construct):
-    expect = ["a", "a", "b", "c", "a"]
-    data = construct(expect)
-    got = cudf.Series(data)
-    assert got.dtype == np.dtype("object")
-    assert len(got) == 5
-    for idx, val in enumerate(expect):
-        assert expect[idx] == got[idx]
-
-
-def test_string_export(ps_gs):
-    ps, gs = ps_gs
-
-    expect = ps
-    got = gs.to_pandas()
-    assert_eq(expect, got)
-
-    expect = np.array(ps)
-    got = gs.to_numpy()
-    assert_eq(expect, got)
-
-    expect = pa.Array.from_pandas(ps)
-    got = gs.to_arrow()
-
-    assert pa.Array.equals(expect, got)
-
-
-@pytest.mark.parametrize(
-    "item",
-    [
-        0,
-        2,
-        4,
-        slice(1, 3),
-        [1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-        [0, 1, 2, 3, 4, 4, 3, 2, 1, 0],
-        np.array([0, 1, 2, 3, 4]),
-        cupy.asarray(np.array([0, 1, 2, 3, 4])),
-    ],
-)
-def test_string_get_item(ps_gs, item):
-    ps, gs = ps_gs
-
-    got = gs.iloc[item]
-    if isinstance(got, cudf.Series):
-        got = got.to_arrow()
-
-    if isinstance(item, cupy.ndarray):
-        item = cupy.asnumpy(item)
-
-    expect = ps.iloc[item]
-    if isinstance(expect, pd.Series):
-        expect = pa.Array.from_pandas(expect)
-        pa.Array.equals(expect, got)
-    else:
-        if got is cudf.NA and expect is None:
-            return
-        assert expect == got
-
-
-@pytest.mark.parametrize(
-    "item",
-    [
-        [True] * 5,
-        [False] * 5,
-        np.array([True] * 5),
-        np.array([False] * 5),
-        cupy.asarray(np.array([True] * 5)),
-        cupy.asarray(np.array([False] * 5)),
-        np.random.default_rng(seed=0)
-        .integers(0, 2, 5)
-        .astype("bool")
-        .tolist(),
-        np.random.default_rng(seed=0).integers(0, 2, 5).astype("bool"),
-        cupy.asarray(
-            np.random.default_rng(seed=0).integers(0, 2, 5).astype("bool")
-        ),
-    ],
-)
-def test_string_bool_mask(ps_gs, item):
-    ps, gs = ps_gs
-
-    got = gs.iloc[item]
-    if isinstance(got, cudf.Series):
-        got = got.to_arrow()
-
-    if isinstance(item, cupy.ndarray):
-        item = cupy.asnumpy(item)
-
-    expect = ps[item]
-    if isinstance(expect, pd.Series):
-        expect = pa.Array.from_pandas(expect)
-        pa.Array.equals(expect, got)
-    else:
-        assert expect == got
-
-
-@pytest.mark.parametrize("item", [0, slice(1, 3), slice(5)])
-def test_string_repr(ps_gs, item):
-    ps, gs = ps_gs
-
-    got_out = gs.iloc[item]
-    expect_out = ps.iloc[item]
-
-    expect = str(expect_out)
-    got = str(got_out)
-
-    if got_out is not cudf.NA and len(got_out) > 1:
-        expect = expect.replace("None", "<NA>")
-
-    assert expect == got or (expect == "None" and got == "<NA>")
-
-
-@pytest.mark.parametrize(
-    "dtype", NUMERIC_TYPES + DATETIME_TYPES + ["bool", "object", "str"]
-)
-def test_string_astype(dtype):
-    if (
-        dtype.startswith("int")
-        or dtype.startswith("uint")
-        or dtype.startswith("long")
-    ):
-        data = ["1", "2", "3", "4", "5"]
-    elif dtype.startswith("float"):
-        data = [
-            "1.0",
-            "2.0",
-            "3.0",
-            "4.0",
-            None,
-            "5.0",
-            "nan",
-            "-INF",
-            "NaN",
-            "inF",
-            "NAn",
-        ]
-    elif dtype.startswith("bool"):
-        data = ["True", "False", "True", "False", "False"]
-    elif dtype.startswith("datetime64"):
-        data = [
-            "2019-06-04T00:00:00",
-            "2019-06-04T12:12:12",
-            "2019-06-03T00:00:00",
-            "2019-05-04T00:00:00",
-            "2018-06-04T00:00:00",
-            "1922-07-21T01:02:03",
-        ]
-    elif dtype == "str" or dtype == "object":
-        data = ["ab", "cd", "ef", "gh", "ij"]
-    ps = pd.Series(data)
-    gs = cudf.Series(data)
-
-    expect = ps.astype(dtype)
-    got = gs.astype(dtype)
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize(
-    "data, scale, precision",
-    [
-        (["1.11", "2.22", "3.33"], 2, 3),
-        (["111", "222", "33"], 0, 3),
-        (["111000", "22000", "3000"], -3, 3),
-        ([None, None, None], 0, 5),
-        ([None, "-2345", None], 0, 5),
-        ([], 0, 5),
-    ],
-)
-@pytest.mark.parametrize(
-    "decimal_dtype",
-    [cudf.Decimal128Dtype, cudf.Decimal64Dtype, cudf.Decimal32Dtype],
-)
-def test_string_to_decimal(data, scale, precision, decimal_dtype):
-    gs = cudf.Series(data, dtype="str")
-    fp = gs.astype(decimal_dtype(scale=scale, precision=precision))
-    got = fp.astype("str")
-    assert_eq(gs, got)
-
-
-def test_string_empty_to_decimal():
-    gs = cudf.Series(["", "-85", ""], dtype="str")
-    got = gs.astype(cudf.Decimal64Dtype(scale=0, precision=5))
-    expected = cudf.Series(
-        [0, -85, 0],
-        dtype=cudf.Decimal64Dtype(scale=0, precision=5),
-    )
-    assert_eq(expected, got)
-
-
-@pytest.mark.parametrize(
-    "data, scale, precision",
-    [
-        (["1.23", "-2.34", "3.45"], 2, 3),
-        (["123", "-234", "345"], 0, 3),
-        (["12300", "-400", "5000.0"], -2, 5),
-        ([None, None, None], 0, 5),
-        ([None, "-100", None], 0, 5),
-        ([], 0, 5),
-    ],
-)
-@pytest.mark.parametrize(
-    "decimal_dtype",
-    [cudf.Decimal128Dtype, cudf.Decimal32Dtype, cudf.Decimal64Dtype],
-)
-def test_string_from_decimal(data, scale, precision, decimal_dtype):
-    decimal_data = []
-    for d in data:
-        if d is None:
-            decimal_data.append(None)
-        else:
-            decimal_data.append(Decimal(d))
-    fp = cudf.Series(
-        decimal_data,
-        dtype=decimal_dtype(scale=scale, precision=precision),
-    )
-    gs = fp.astype("str")
-    got = gs.astype(decimal_dtype(scale=scale, precision=precision))
-    assert_eq(fp, got)
-
-
-@pytest.mark.parametrize(
-    "dtype", NUMERIC_TYPES + DATETIME_TYPES + ["bool", "object", "str"]
-)
-def test_string_empty_astype(dtype):
-    data = []
-    ps = pd.Series(data, dtype="str")
-    gs = cudf.Series(data, dtype="str")
-
-    expect = ps.astype(dtype)
-    got = gs.astype(dtype)
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize("dtype", NUMERIC_TYPES + DATETIME_TYPES + ["bool"])
-def test_string_numeric_astype(dtype):
-    if dtype.startswith("bool"):
-        data = [1, 0, 1, 0, 1]
-    elif (
-        dtype.startswith("int")
-        or dtype.startswith("uint")
-        or dtype.startswith("long")
-    ):
-        data = [1, 2, 3, 4, 5]
-    elif dtype.startswith("float"):
-        data = [1.0, 2.0, 3.0, 4.0, 5.0]
-    elif dtype.startswith("datetime64"):
-        # pandas rounds the output format based on the data
-        # Use numpy instead
-        # but fix '2011-01-01T00:00:00' -> '2011-01-01 00:00:00'
-        data = [1000000001, 2000000001, 3000000001, 4000000001, 5000000001]
-        ps = np.asarray(data, dtype=dtype).astype(str)
-        ps = np.array([i.replace("T", " ") for i in ps])
-
-    if not dtype.startswith("datetime64"):
-        ps = pd.Series(data, dtype=dtype)
-
-    gs = cudf.Series(data, dtype=dtype)
-
-    expect = pd.Series(ps.astype("str"))
-    got = gs.astype("str")
-
-    assert_eq(expect, got)
-
-
-@pytest.mark.parametrize("dtype", NUMERIC_TYPES + DATETIME_TYPES + ["bool"])
-def test_string_empty_numeric_astype(dtype):
-    data = []
-
-    if dtype.startswith("datetime64"):
-        ps = pd.Series(data, dtype="datetime64[ns]")
-    else:
-        ps = pd.Series(data, dtype=dtype)
-    gs = cudf.Series(data, dtype=dtype)
-
-    expect = ps.astype("str")
-    got = gs.astype("str")
-
-    assert_eq(expect, got)
 
 
 @pytest.mark.parametrize("ascending", [True, False])
